demo of the current functionality (#17)
* make sure we always have float64 coordinates

* expose `infer_cell_geometries`

* prepare for notebooks

* demo notebook on `infer_*`

The `visualize_grid` utility may be moved elsewhere at some point.

* demo of the overlap query

+++ python/grid_indexing/grids.py
@@ -22,7 +22,7 @@ def as_components(boundaries):
     geom_offsets = np.arange(np.prod(vertices.shape[:-2]) + 1, dtype="int32")
     ring_offsets = geom_offsets * coords_per_pixel
 
-    return coords, geom_offsets, ring_offsets
+    return coords.astype("float64"), geom_offsets, ring_offsets
 
 
 def infer_grid_type(ds: xr.Dataset):
